Module Avutil

Common code shared across all FFmpeg libraries.

Line

type input

type output

Container

type 'a container

Media types

type audio = [

| `Audio

]

type video = [

| `Video

]

type subtitle = [

| `Subtitle

]

type media_type = [

| audio

| video

| subtitle

| `Unknown

| `Data

| `Attachment

| `Nb

]

Format

type ('line, 'media) format

Frame

type 'media frame

val frame_pts : _ frame -> Stdlib.Int64.t option

Avutil.frame_pts frame returns the presentation timestamp in time_base units (time when frame should be shown to user).

val frame_set_pts : _ frame -> Stdlib.Int64.t option -> unit

Avutil.frame_set_pts frame pts sets the presentation time for this frame.

val frame_best_effort_timestamp : _ frame -> Stdlib.Int64.t option

Avutil.frame_best_effort_timestamp frame returns the frame timestamp estimated using various heuristics, in stream time base

val frame_pkt_duration : _ frame -> Stdlib.Int64.t option

duration of the corresponding packet, expressed in AVStream->time_base units.

val frame_copy : 'a frame -> 'b frame -> unit

Avutil.frame_copy src dst copies data from src into dst

Exception

type error = [

| `Bsf_not_found

| `Decoder_not_found

| `Demuxer_not_found

| `Encoder_not_found

| `Eof

| `Exit

| `Filter_not_found

| `Invalid_data

| `Muxer_not_found

| `Option_not_found

| `Patch_welcome

| `Protocol_not_found

| `Stream_not_found

| `Bug

| `Eagain

| `Unknown

| `Experimental

| `Other of int

| `Failure of string

]

Internal errors.

exception Error of error

val string_of_error : error -> string

type data = (int, Stdlib.Bigarray.int8_unsigned_elt, Stdlib.Bigarray.c_layout) Stdlib.Bigarray.Array1.t

val create_data : int -> data

type rational = {

num : int;

den : int;

}

val string_of_rational : rational -> string

Constants

val qp2lambda : int

Timestamp

module Time_format : sig ... end

Formats for time.

val time_base : unit -> rational

Return the time base of FFmpeg.

Logging utilities

module Log : sig ... end

Audio utilities

module Channel_layout : sig ... end

Formats for channels layouts.

module Sample_format : sig ... end

Formats for audio samples.

Video utilities

module Pixel_format : sig ... end

Formats for pixels.

module Audio : sig ... end

module Video : sig ... end

Subtitle utilities

module Subtitle : sig ... end

module Options : sig ... end

type value = [

| `String of string

| `Int of int

| `Int64 of int64

| `Float of float

]

type opts = (string, value) Stdlib.Hashtbl.t

val opts_default : opts option -> opts

val mk_opts_array : opts -> (string * string) array

val string_of_opts : opts -> string

val mk_audio_opts : ?⁠opts:opts -> ?⁠channels:int -> ?⁠channel_layout:Channel_layout.t -> sample_rate:int -> sample_format:Sample_format.t -> time_base:rational -> unit -> opts

val mk_video_opts : ?⁠opts:opts -> ?⁠frame_rate:rational -> pixel_format:Pixel_format.t -> width:int -> height:int -> time_base:rational -> unit -> opts

val filter_opts : string array -> opts -> unit

HwContext

module HwContext : sig ... end