Module Avutil.Audio

val create_frame : Sample_format.t -> Channel_layout.t -> int -> int -> audio frame

Avutil.Audio.create_frame sample_format channel_layout sample_rate samples allocates a new audio frame.

val frame_get_sample_format : audio frame -> Sample_format.t

Avutil.Audio.frame_get_sample_format frame returns the sample format of the current frame.

val frame_get_sample_rate : audio frame -> int

Avutil.Audio.frame_get_sample_rate frame returns the sample rate of the current frame.

val frame_get_channels : audio frame -> int

Avutil.Audio.frame_get_channels frame returns the number of audio channels in the current frame.

val frame_get_channel_layout : audio frame -> Channel_layout.t

Avutil.Audio.frame_get_channel_layout frame returns the channel layout for the current frame.

val frame_nb_samples : audio frame -> int

Avutil.Audio.frame_nb_samples frame returns the number of audio samples per channel in the current frame.

val frame_copy_samples : audio frame -> int -> audio frame -> int -> int -> unit

Abutil.Audio.frame_copy_samples src src_offset dst dst_offset len copies len samples from src starting at position src_offset into dst starting at position dst_offset.