Module Avutil.Channel_layout

Formats for channels layouts.

type layout = Channel_layout.t

Channel layout formats.

type t

val standard_layouts : t list

List of standard channel layouts.

val stereo : t

val mono : t

val five_point_one : t

val compare : t -> t -> bool

Compare two channel layouts.

val find : string -> t

Return a channel layout id that matches name. Raises Not_found otherwise. name can be one or several of the following notations, separated by '+' or '|':

the name of an usual channel layout (mono, stereo, 4.0, quad, 5.0, 5.0(side), 5.1, 5.1(side), 7.1, 7.1(wide), downmix);

the name of a single channel (FL, FR, FC, LFE, BL, BR, FLC, FRC, BC, SL, SR, TC, TFL, TFC, TFR, TBL, TBC, TBR, DL, DR);

a number of channels, in decimal, optionally followed by 'c', yielding the default channel layout for that number of channels;

a channel layout mask, in hexadecimal starting with "0x" (see the AV_CH_* macros).

val get_description : t -> string

Return a description of the channel layout.

val get_nb_channels : t -> int

Return the number of channels in the channel layout.

val get_default : int -> t

Return default channel layout for a given number of channels. Raises Not_found if not found.

val get_mask : t -> int64 option

Return a native channel layout mask, suitable for filters channel_layout.

val get_native_id : t -> int64 option

Deprecated