Module Avutil.Channel_layout

Formats for channels layouts.

type t = Channel_layout.t

Channel layout formats.

val find : string -> t

Return a channel layout id that matches name. Raises Not_found otherwise. name can be one or several of the following notations, separated by '+' or '|':

the name of an usual channel layout (mono, stereo, 4.0, quad, 5.0, 5.0(side), 5.1, 5.1(side), 7.1, 7.1(wide), downmix);

the name of a single channel (FL, FR, FC, LFE, BL, BR, FLC, FRC, BC, SL, SR, TC, TFL, TFC, TFR, TBL, TBC, TBR, DL, DR);

a number of channels, in decimal, optionally followed by 'c', yielding the default channel layout for that number of channels;

a channel layout mask, in hexadecimal starting with "0x" (see the AV_CH_* macros).

val get_description : ?⁠channels:int -> t -> string

Return a description of the channel layout.

val get_nb_channels : t -> int

Return the number of channels in the channel layout.

val get_default : int -> t

Return default channel layout for a given number of channels. Raises Not_found if not found.

val get_id : t -> int64

Return the internal ID for a channel layout. This number should be passed as the "channel_layout" opts in Av.new_audio_stream .