Module Avutil.Frame

type 'media t

val pts : _ t -> Stdlib.Int64.t option

Avutil.Frame.pts frame returns the presentation timestamp in time_base units (time when frame should be shown to user).

val set_pts : _ t -> Stdlib.Int64.t option -> unit

Avutil.Frame.set_pts frame pts sets the presentation time for this frame.

val duration : _ t -> Stdlib.Int64.t option

Avutil.Frame.duration frame returns the frame duration in time_base, when available.

val set_duration : _ t -> Stdlib.Int64.t option -> unit

Avutil.Frame.set_duration frame duration sets the frame duration.

val pkt_dts : _ t -> Stdlib.Int64.t option

Avutil.Frame.pkt_dts frame returns DTS copied from the AVPacket that triggered returning this frame.

val set_pkt_dts : _ t -> Stdlib.Int64.t option -> unit

Avutil.Frame.set_pkt_dts frame dts sets pkt_dts value for this frame.

val metadata : _ t -> (string * string) list

Avutil.Frame.metadata frame returns the frame's metadata.

val set_metadata : _ t -> (string * string) list -> unit

Avutil.Frame.set_metadata frame metadata sets the frame's metadata.

val best_effort_timestamp : _ t -> Stdlib.Int64.t option

Avutil.frame_best_effort_timestamp frame returns the frame timestamp estimated using various heuristics, in stream time base

val copy : 'a t -> 'b t -> unit

Avutil.frame_copy src dst copies data from src into dst