Module Avutil.Pixel_format

Formats for pixels.

type t = Pixel_format.t

Pixels formats.

type flag = Avutil__.Pixel_format_flag.t

Pixel format flags.

type component_descriptor = {

plane : int;

step : int;

offset : int;

shift : int;

depth : int;

}

Pixel format component descriptor

type descriptor = private {

name : string;

nb_components : int;

log2_chroma_w : int;

log2_chroma_h : int;

flags : flag list;

comp : component_descriptor list;

alias : string option;

}

Pixel format descriptor.

val descriptor : t -> descriptor

Return the pixel's format descriptor. Raises Not_found if descriptor could not be found.

val bits : descriptor -> int

Return the number of bits of the pixel format.

val planes : t -> int

Return the number of planes of the pixel format.

val to_string : t -> string option

Pixel_format.to_string f Return a string representation of the pixel format f if it exists

val of_string : string -> t

Pixel_format.of_string s Convert the string s into a Pixel_format.t. Raises Error if s is not a valid format.

val get_id : t -> int

Return the internal ID of the pixel format.

val find_id : int -> t

Find a sample pixel from its ID. Raises Not_found when none exist.