Module Avutil.Pixel_format

Formats for pixels.

type t = [

| `None

| `Yuv420p

| `Yuyv422

| `Rgb24

| `Bgr24

| `Yuv422p

| `Yuv444p

| `Yuv410p

| `Yuv411p

| `Gray8

| `Monowhite

| `Monoblack

| `Pal8

| `Yuvj420p

| `Yuvj422p

| `Yuvj444p

| `Uyvy422

| `Uyyvyy411

| `Bgr8

| `Bgr4

| `Bgr4_byte

| `Rgb8

| `Rgb4

| `Rgb4_byte

| `Nv12

| `Nv21

| `Argb

| `Rgba

| `Abgr

| `Bgra

| `Gray16be

| `Gray16le

| `Yuv440p

| `Yuvj440p

| `Yuva420p

| `Rgb48be

| `Rgb48le

| `Rgb565be

| `Rgb565le

| `Rgb555be

| `Rgb555le

| `Bgr565be

| `Bgr565le

| `Bgr555be

| `Bgr555le

| `Vaapi_moco

| `Vaapi_idct

| `Vaapi_vld

| `Vaapi

| `Yuv420p16le

| `Yuv420p16be

| `Yuv422p16le

| `Yuv422p16be

| `Yuv444p16le

| `Yuv444p16be

| `Dxva2_vld

| `Rgb444le

| `Rgb444be

| `Bgr444le

| `Bgr444be

| `Ya8

| `Y400a

| `Gray8a

| `Bgr48be

| `Bgr48le

| `Yuv420p9be

| `Yuv420p9le

| `Yuv420p10be

| `Yuv420p10le

| `Yuv422p10be

| `Yuv422p10le

| `Yuv444p9be

| `Yuv444p9le

| `Yuv444p10be

| `Yuv444p10le

| `Yuv422p9be

| `Yuv422p9le

| `Gbrp

| `Gbr24p

| `Gbrp9be

| `Gbrp9le

| `Gbrp10be

| `Gbrp10le

| `Gbrp16be

| `Gbrp16le

| `Yuva422p

| `Yuva444p

| `Yuva420p9be

| `Yuva420p9le

| `Yuva422p9be

| `Yuva422p9le

| `Yuva444p9be

| `Yuva444p9le

| `Yuva420p10be

| `Yuva420p10le

| `Yuva422p10be

| `Yuva422p10le

| `Yuva444p10be

| `Yuva444p10le

| `Yuva420p16be

| `Yuva420p16le

| `Yuva422p16be

| `Yuva422p16le

| `Yuva444p16be

| `Yuva444p16le

| `Vdpau

| `Xyz12le

| `Xyz12be

| `Nv16

| `Nv20le

| `Nv20be

| `Rgba64be

| `Rgba64le

| `Bgra64be

| `Bgra64le

| `Yvyu422

| `Ya16be

| `Ya16le

| `Gbrap

| `Gbrap16be

| `Gbrap16le

| `Qsv

| `Mmal

| `D3d11va_vld

| `Cuda

| `_0rgb

| `Rgb0

| `_0bgr

| `Bgr0

| `Yuv420p12be

| `Yuv420p12le

| `Yuv420p14be

| `Yuv420p14le

| `Yuv422p12be

| `Yuv422p12le

| `Yuv422p14be

| `Yuv422p14le

| `Yuv444p12be

| `Yuv444p12le

| `Yuv444p14be

| `Yuv444p14le

| `Gbrp12be

| `Gbrp12le

| `Gbrp14be

| `Gbrp14le

| `Yuvj411p

| `Bayer_bggr8

| `Bayer_rggb8

| `Bayer_gbrg8

| `Bayer_grbg8

| `Bayer_bggr16le

| `Bayer_bggr16be

| `Bayer_rggb16le

| `Bayer_rggb16be

| `Bayer_gbrg16le

| `Bayer_gbrg16be

| `Bayer_grbg16le

| `Bayer_grbg16be

| `Xvmc

| `Yuv440p10le

| `Yuv440p10be

| `Yuv440p12le

| `Yuv440p12be

| `Ayuv64le

| `Ayuv64be

| `Videotoolbox

| `P010le

| `P010be

| `Gbrap12be

| `Gbrap12le

| `Gbrap10be

| `Gbrap10le

| `Mediacodec

| `Gray12be

| `Gray12le

| `Gray10be

| `Gray10le

| `P016le

| `P016be

| `D3d11

| `Gray9be

| `Gray9le

| `Gbrpf32be

| `Gbrpf32le

| `Gbrapf32be

| `Gbrapf32le

| `Drm_prime

| `Opencl

| `Gray14be

| `Gray14le

| `Grayf32be

| `Grayf32le

| `Yuva422p12be

| `Yuva422p12le

| `Yuva444p12be

| `Yuva444p12le

| `Nv24

| `Nv42

| `Vulkan

| `Y210be

| `Y210le

| `X2rgb10le

| `X2rgb10be

]

Pixels formats.

type flag = [

| `Be

| `Pal

| `Bitstream

| `Hwaccel

| `Planar

| `Rgb

| `Pseudopal

| `Alpha

| `Bayer

| `Float

]

Pixel format flags.

type component_descriptor = {

plane : int;

step : int;

offset : int;

shift : int;

depth : int;

}

Pixel format component descriptor

type descriptor = private {

name : string;

nb_components : int;

log2_chroma_w : int;

log2_chroma_h : int;

flags : flag list;

comp : component_descriptor list;

alias : string option;

}

Pixel format descriptor.

val descriptor : t -> descriptor

Return the pixel's format descriptor. Raises Not_found if descriptor could not be found.

val bits : descriptor -> int

Return the number of bits of the pixel format.

val planes : t -> int

Return the number of planes of the pixel format.

val to_string : t -> string option

Pixel_format.to_string f Return a string representation of the pixel format f if it exists

val of_string : string -> t

Pixel_format.of_string s Convert the string s into a Pixel_format.t. Raises Error if s is not a valid format.

val get_id : t -> int

Return the internal ID of the pixel format.

val find_id : int -> t

Find a sample pixel from its ID. Raises Not_found when none exist.