Module Avutil.Subtitle

val time_base : unit -> rational

Return the time base for subtitles.

val create_frame : float -> float -> string list -> subtitle frame

Avutil.Subtitle.create_frame start end lines create a subtitle frame from lines which is displayed at start time and hidden at end time in seconds. Raises Error if the allocation failed.

val frame_to_lines : subtitle frame -> float * float * string list

Convert subtitle frame to lines. The two float are the start and the end dislpay time in seconds.