Module Avutil.Video

type planes = (data * int) array

val create_frame : int -> int -> Pixel_format.t -> video frame

Avutil.Video.create_frame w h pf create a video frame with w width, h height and pf pixel format. Raises Error if the allocation failed.

val frame_get_linesize : video frame -> int -> int

Avutil.Video.frame_get_linesize vf n return the line size of the n plane of the vf video frame. Raises Error if n is out of boundaries.

val frame_visit : make_writable:bool -> (planes -> unit) -> video frame -> video frame

Avutil.Video.frame_visit ~make_writable:wrt f vf call the f function with planes wrapping the vf video frame data. The make_writable:wrt parameter must be set to true if the f function writes in the planes. Access to the frame through the planes is safe as long as it occurs in the f function and the frame is not sent to an encoder. The same frame is returned for convenience. Raises Error if the make frame writable operation failed.

val frame_get_width : video frame -> int

Avutil.Video.frame_get_width frame returns the frame width

val frame_get_height : video frame -> int

Avutil.Video.frame_get_height frame returns the frame height

val frame_get_pixel_format : video frame -> Pixel_format.t

Avutil.Video.frame_get_pixel_format frame returns frame's pixel format.

val frame_get_pixel_aspect : video frame -> rational option

Avutil.Video.frame_get_pixel_aspect frame returns the frame's pixel aspect.

val frame_get_color_space : video frame -> Color_space.t

Avutil.Video.frame_get_color_space frame returns frame's color space

val frame_get_color_range : video frame -> Color_range.t

Avutil.Video.frame_get_color_range frame returns frame's color range

val frame_get_color_primaries : video frame -> Color_primaries.t

val frame_get_color_trc : video frame -> Color_trc.t

val frame_get_chroma_location : video frame -> Chroma_location.t